+++ packages/frontend/core/src/blocksuite/block-suite-editor/blocksuite-editor.tsx
@@ -1,30 +1,50 @@
-import { useRefEffect } from '@affine/component';
 import { EditorLoading } from '@affine/component/page-detail-skeleton';
+import type {
+  EdgelessEditor,
+  PageEditor,
+} from '@affine/core/blocksuite/editors';
 import { ServerService } from '@affine/core/modules/cloud';
 import {
   EditorSettingService,
   fontStyleOptions,
 } from '@affine/core/modules/editor-setting';
+import { FeatureFlagService } from '@affine/core/modules/feature-flag';
 import {
   customImageProxyMiddleware,
   ImageProxyService,
 } from '@blocksuite/affine/blocks/image';
+import { appendParagraphCommand } from '@blocksuite/affine/blocks/paragraph';
+import type { DocTitle } from '@blocksuite/affine/fragments/doc-title';
 import { DisposableGroup } from '@blocksuite/affine/global/disposable';
-import type { DocMode } from '@blocksuite/affine/model';
+import type { DocMode, RootBlockModel } from '@blocksuite/affine/model';
+import { focusBlockEnd } from '@blocksuite/affine/shared/commands';
 import { LinkPreviewerService } from '@blocksuite/affine/shared/services';
+import { getLastNoteBlock } from '@blocksuite/affine/shared/utils';
+import type { BlockStdScope, EditorHost } from '@blocksuite/affine/std';
 import type { Store } from '@blocksuite/affine/store';
 import { Slot } from '@radix-ui/react-slot';
 import { useLiveData, useService } from '@toeverything/infra';
 import { cssVar } from '@toeverything/theme';
+import clsx from 'clsx';
 import type { CSSProperties, HTMLAttributes } from 'react';
-import { useCallback, useEffect, useMemo, useState } from 'react';
+import { useCallback, useEffect, useMemo, useRef, useState } from 'react';
 
 import type { DefaultOpenProperty } from '../../components/doc-properties';
-import {
-  type AffineEditorContainer,
-  BlocksuiteEditorContainer,
-} from './blocksuite-editor-container';
+import { BlocksuiteDocEditor, BlocksuiteEdgelessEditor } from './lit-adaper';
 import { NoPageRootError } from './no-page-error';
+import * as styles from './styles.css';
+
+export interface AffineEditorContainer extends HTMLElement {
+  page: Store;
+  doc: Store;
+  docTitle: DocTitle;
+  host?: EditorHost;
+  model: RootBlockModel | null;
+  updateComplete: Promise<boolean>;
+  mode: DocMode;
+  origin: HTMLDivElement;
+  std: BlockStdScope;
+}
 
 export interface EditorProps extends HTMLAttributes<HTMLDivElement> {
   page: Store;
@@ -47,6 +67,113 @@ const BlockSuiteEditorImpl = ({
   defaultOpenProperty,
   ...props
 }: EditorProps) => {
+  const rootRef = useRef<HTMLDivElement>(null);
+  const docRef = useRef<PageEditor>(null);
+  const docTitleRef = useRef<DocTitle>(null);
+  const edgelessRef = useRef<EdgelessEditor>(null);
+  const featureFlags = useService(FeatureFlagService).flags;
+  const enableEditorRTL = useLiveData(featureFlags.enable_editor_rtl.$);
+  const editorSetting = useService(EditorSettingService).editorSetting;
+  const server = useService(ServerService).server;
+
+  const { enableMiddleClickPaste } = useLiveData(
+    editorSetting.settings$.selector(s => ({
+      enableMiddleClickPaste: s.enableMiddleClickPaste,
+    }))
+  );
+
+  /**
+   * mimic an AffineEditorContainer using proxy
+   */
+  const affineEditorContainerProxy = useMemo(() => {
+    const api = {
+      get page() {
+        return page;
+      },
+      get doc() {
+        return page;
+      },
+      get docTitle() {
+        return docTitleRef.current;
+      },
+      get host() {
+        return (
+          (mode === 'page'
+            ? docRef.current?.host
+            : edgelessRef.current?.host) ?? null
+        );
+      },
+      get model() {
+        return page.root as any;
+      },
+      get updateComplete() {
+        return mode === 'page'
+          ? docRef.current?.updateComplete
+          : edgelessRef.current?.updateComplete;
+      },
+      get mode() {
+        return mode;
+      },
+      get origin() {
+        return rootRef.current;
+      },
+      get std() {
+        return mode === 'page' ? docRef.current?.std : edgelessRef.current?.std;
+      },
+    };
+
+    const proxy = new Proxy(api, {
+      has(_, prop) {
+        return (
+          Reflect.has(api, prop) ||
+          (rootRef.current ? Reflect.has(rootRef.current, prop) : false)
+        );
+      },
+      get(_, prop) {
+        if (Reflect.has(api, prop)) {
+          return api[prop as keyof typeof api];
+        }
+        if (rootRef.current && Reflect.has(rootRef.current, prop)) {
+          const maybeFn = Reflect.get(rootRef.current, prop);
+          if (typeof maybeFn === 'function') {
+            return maybeFn.bind(rootRef.current);
+          } else {
+            return maybeFn;
+          }
+        }
+        return undefined;
+      },
+    }) as AffineEditorContainer;
+
+    return proxy;
+  }, [mode, page]);
+
+  const handleClickPageModeBlank = useCallback(() => {
+    if (shared || readonly || page.readonly) return;
+    const std = affineEditorContainerProxy.host?.std;
+    if (!std) {
+      return;
+    }
+    const note = getLastNoteBlock(page);
+    if (note) {
+      const lastBlock = note.lastChild();
+      if (
+        lastBlock &&
+        lastBlock.flavour === 'affine:paragraph' &&
+        lastBlock.text?.length === 0
+      ) {
+        const focusBlock = std.view.getBlock(lastBlock.id) ?? undefined;
+        std.command.exec(focusBlockEnd, {
+          focusBlock,
+          force: true,
+        });
+        return;
+      }
+    }
+
+    std.command.exec(appendParagraphCommand);
+  }, [affineEditorContainerProxy.host?.std, page, readonly, shared]);
+
   useEffect(() => {
     const disposable = page.slots.blockUpdated.subscribe(() => {
       disposable.unsubscribe();
@@ -59,67 +186,104 @@ const BlockSuiteEditorImpl = ({
     };
   }, [page]);
 
-  const server = useService(ServerService).server;
-
-  const editorRef = useRefEffect(
-    (editor: AffineEditorContainer) => {
-      globalThis.currentEditor = editor;
-      let canceled = false;
-      const disposableGroup = new DisposableGroup();
-
-      // Invoke onLoad once the editor has been mounted to the DOM.
-      if (canceled) {
-        return;
-      }
-
-      // provide image proxy endpoint to blocksuite
-      const imageProxyUrl = new URL(
-        BUILD_CONFIG.imageProxyUrl,
-        server.baseUrl
-      ).toString();
-
-      const linkPreviewUrl = new URL(
-        BUILD_CONFIG.linkPreviewUrl,
-        server.baseUrl
-      ).toString();
-
-      editor.std.clipboard.use(customImageProxyMiddleware(imageProxyUrl));
-      page.get(LinkPreviewerService).setEndpoint(linkPreviewUrl);
-      page.get(ImageProxyService).setImageProxyURL(imageProxyUrl);
-
-      editor.updateComplete
-        .then(() => {
-          if (onEditorReady) {
-            const dispose = onEditorReady(editor);
-            if (dispose) {
-              disposableGroup.add(dispose);
-            }
-          }
-        })
-        .catch(error => {
-          console.error('Error updating editor', error);
-        });
-
-      return () => {
-        canceled = true;
-        disposableGroup.dispose();
+  useEffect(() => {
+    const editorContainer = rootRef.current;
+    if (editorContainer) {
+      const handleMiddleClick = (e: MouseEvent) => {
+        if (!enableMiddleClickPaste && e.button === 1) {
+          e.preventDefault();
+        }
       };
-    },
-    [onEditorReady, page, server]
-  );
+      editorContainer.addEventListener('pointerup', handleMiddleClick, {
+        capture: true,
+      });
+      editorContainer.addEventListener('auxclick', handleMiddleClick, {
+        capture: true,
+      });
+      return () => {
+        editorContainer?.removeEventListener('pointerup', handleMiddleClick, {
+          capture: true,
+        });
+        editorContainer?.removeEventListener('auxclick', handleMiddleClick, {
+          capture: true,
+        });
+      };
+    }
+    return;
+  }, [enableMiddleClickPaste]);
+
+  useEffect(() => {
+    const editor = affineEditorContainerProxy;
+    globalThis.currentEditor = editor;
+    const disposableGroup = new DisposableGroup();
+    let canceled = false;
+
+    // provide image proxy endpoint to blocksuite
+    const imageProxyUrl = new URL(
+      BUILD_CONFIG.imageProxyUrl,
+      server.baseUrl
+    ).toString();
+
+    const linkPreviewUrl = new URL(
+      BUILD_CONFIG.linkPreviewUrl,
+      server.baseUrl
+    ).toString();
+
+    editor.std.clipboard.use(customImageProxyMiddleware(imageProxyUrl));
+    page.get(LinkPreviewerService).setEndpoint(linkPreviewUrl);
+    page.get(ImageProxyService).setImageProxyURL(imageProxyUrl);
+
+    editor.updateComplete
+      .then(() => {
+        if (onEditorReady && !canceled) {
+          const dispose = onEditorReady(editor);
+          if (dispose) {
+            disposableGroup.add(dispose);
+          }
+        }
+      })
+      .catch(error => {
+        console.error('Error updating editor', error);
+      });
+
+    return () => {
+      canceled = true;
+      disposableGroup.dispose();
+    };
+  }, [affineEditorContainerProxy, onEditorReady, page, server]);
 
   return (
-    <BlocksuiteEditorContainer
+    <div
       {...props}
-      mode={mode}
-      page={page}
-      shared={shared}
-      readonly={readonly}
-      defaultOpenProperty={defaultOpenProperty}
-      ref={editorRef}
-      className={className}
+      data-testid={`editor-${page.id}`}
+      dir={enableEditorRTL ? 'rtl' : 'ltr'}
+      className={clsx(
+        `editor-wrapper ${mode}-mode`,
+        styles.docEditorRoot,
+        className
+      )}
       style={style}
-    />
+      data-affine-editor-container
+      ref={rootRef}
+    >
+      {mode === 'page' ? (
+        <BlocksuiteDocEditor
+          shared={shared}
+          page={page}
+          ref={docRef}
+          readonly={readonly}
+          titleRef={docTitleRef}
+          onClickBlank={handleClickPageModeBlank}
+          defaultOpenProperty={defaultOpenProperty}
+        />
+      ) : (
+        <BlocksuiteEdgelessEditor
+          shared={shared}
+          page={page}
+          ref={edgelessRef}
+        />
+      )}
+    </div>
   );
 };
 
@@ -133,7 +297,6 @@ export const BlockSuiteEditor = (props: EditorProps) => {
       fontFamily: s.fontFamily,
       customFontFamily: s.customFontFamily,
       fullWidthLayout: s.fullWidthLayout,
-      disableMiddleClickPaste: s.disableMiddleClickPaste,
     }))
   );
   const fontFamily = useMemo(() => {
@@ -170,24 +333,12 @@ export const BlockSuiteEditor = (props: EditorProps) => {
     };
   }, [props.page]);
 
-  const handleMouseDown = useCallback(
-    (e: React.MouseEvent<HTMLDivElement>) => {
-      if (settings.disableMiddleClickPaste && e.button === 1) {
-        e.preventDefault();
-      }
-    },
-    [settings.disableMiddleClickPaste]
-  );
-
   if (error) {
     throw error;
   }
 
   return (
-    <Slot
-      style={{ '--affine-font-family': fontFamily } as CSSProperties}
-      onMouseDown={handleMouseDown}
-    >
+    <Slot style={{ '--affine-font-family': fontFamily } as CSSProperties}>
       {isLoading ? (
         <EditorLoading />
       ) : (
