+++ packages/frontend/core/src/modules/editor-setting/schema.ts
@@ -35,7 +35,7 @@ const AffineEditorSettingSchema = z.object({
     ])
     .default('open-in-active-view'),
   // linux only:
-  disableMiddleClickPaste: z.boolean().default(false),
+  enableMiddleClickPaste: z.boolean().default(false),
 });
 
 export const EditorSettingSchema = BSEditorSettingSchema.merge(
